cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.25.1
pytest==6.2.2
google-cloud-storage>=1.26.0, <2.0.0dev
google-cloud-aiplatform==0.7.1
tensorflow==2.4.1
pillow==8.2.0